backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
requests==2.31.0
eventlet==0.33.3
